My group project for ITSC-3155 at Charlotte
This is the final project for my software engineering course at the University of North Carolina at Charlotte. We where given 3 weeks to make a working product using FastAPI and Pycharm. We where unfortunately unable to flesh out the rest of the development process due to a lack of time in the given course.